Software/rover_sim/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rover_sim"
version = "0.0.1"
description = "Kinematic rover simulator and shared interfaces for the PolarRover drive stack"
requires-python = ">=3.10"
authors = [{ name = "hoffmaao" }]
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["rover_sim"]
